+++ packages/frontend/src/components/Explorer/ResultsCard/ColumnHeaderContextMenu.styles.tsx
@@ -1,0 +1,10 @@
+import { Button } from '@blueprintjs/core';
+import styled from 'styled-components';
+
+export const BolderLabel = styled.span`
+    font-weight: 600;
+`;
+
+export const FlatButton = styled(Button)`
+    min-height: 16px !important;
+`;