+++ packages/frontend/src/hooks/cartesianChartConfig/useCartesianChartConfig.test.ts
@@ -1,5 +1,20 @@
-import { sortDimensions } from './useCartesianChartConfig';
-import { explore } from './useCartesianChartConfig.mock';
+import {
+    existingMixedSeries,
+    expectedMixedSeriesMap,
+    expectedPivotedSeriesMap,
+    expectedSimpleSeriesMap,
+    explore,
+    groupedMixedSeries,
+    mergedMixedSeries,
+    pivotSeriesMapArgs,
+    simpleSeriesMapArgs,
+} from './useCartesianChartConfig.mock';
+import {
+    getExpectedSeriesMap,
+    getSeriesGroupedByField,
+    mergeExistingAndExpectedSeries,
+    sortDimensions,
+} from './utils';
 
 describe('sortDimensions', () => {
     test('should not sort anything if no explore', () => {
@@ -98,5 +113,59 @@ describe('sortDimensions', () => {
             'dimension_timestamp',
             'dimension_string',
         ]);
+    });
+});
+
+describe('getExpectedSeriesMap', () => {
+    test('should return series without pivot', () => {
+        expect(getExpectedSeriesMap(simpleSeriesMapArgs)).toStrictEqual(
+            expectedSimpleSeriesMap,
+        );
+    });
+    test('should return series with pivot', () => {
+        expect(getExpectedSeriesMap(pivotSeriesMapArgs)).toStrictEqual(
+            expectedPivotedSeriesMap,
+        );
+    });
+});
+
+describe('mergeExistingAndExpectedSeries', () => {
+    test('should return empty list when expected series is empty', () => {
+        expect(
+            mergeExistingAndExpectedSeries({
+                expectedSeriesMap: {},
+                existingSeries: Object.values(expectedSimpleSeriesMap),
+            }),
+        ).toStrictEqual([]);
+    });
+    test('should return all expected series when existing series is empty', () => {
+        expect(
+            mergeExistingAndExpectedSeries({
+                expectedSeriesMap: expectedSimpleSeriesMap,
+                existingSeries: [],
+            }),
+        ).toStrictEqual(Object.values(expectedSimpleSeriesMap));
+        expect(
+            mergeExistingAndExpectedSeries({
+                expectedSeriesMap: expectedPivotedSeriesMap,
+                existingSeries: [],
+            }),
+        ).toStrictEqual(Object.values(expectedPivotedSeriesMap));
+    });
+    test('should return new series in correct order', () => {
+        expect(
+            mergeExistingAndExpectedSeries({
+                expectedSeriesMap: expectedMixedSeriesMap,
+                existingSeries: existingMixedSeries,
+            }),
+        ).toStrictEqual(Object.values(mergedMixedSeries));
+    });
+});
+
+describe('getSeriesGroupedByField', () => {
+    test('should return series grouped by Y field', () => {
+        expect(
+            getSeriesGroupedByField(Object.values(mergedMixedSeries)),
+        ).toStrictEqual(groupedMixedSeries);
     });
 });