+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# [0.255.0](https://github.com/lightdash/lightdash/compare/0.254.0...0.255.0) (2022-09-16)
+
+
+### Features
+
+* results table sort limit pills in view mode ([#3251](https://github.com/lightdash/lightdash/issues/3251)) ([1c03457](https://github.com/lightdash/lightdash/commit/1c0345773885d946b32b946866bacb85a0f10327))
+
+# [0.254.0](https://github.com/lightdash/lightdash/compare/0.253.0...0.254.0) (2022-09-15)
+
+
+### Bug Fixes
+
+* sort table calculations ([#3243](https://github.com/lightdash/lightdash/issues/3243)) ([0157322](https://github.com/lightdash/lightdash/commit/0157322c66cd2039b9d990b0d9627109d403b1db))
+
+
+### Features
+
+* order chart series ([#3206](https://github.com/lightdash/lightdash/issues/3206)) ([a399f40](https://github.com/lightdash/lightdash/commit/a399f4039b4b3cd43337eed09811dc2552ced320))
+
 # [0.253.0](https://github.com/lightdash/lightdash/compare/0.252.2...0.253.0) (2022-09-15)